requirements.txt
dvc>=3.39.0
dvc-gs>=3.0.1
hydra-core>=1.3.2
numpy>=1.24.3
omegaconf>=2.3.0
pandas>=2.1.4
pytorch_lightning>=2.1.3
scikit_learn>=1.2.2
torch>=2.1.0
transformers>=4.32.1
wandb>=0.16.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.6
google-cloud-storage